requirements.txt
django==1.8.2
psycopg2==2.5.3
py-bcrypt==0.4
python-memcached==1.53

# Debug Tools
sqlparse==0.1.6
django-debug-toolbar==1.2.1
django-devserver==0.8.0

# AWS
django-storages==1.1.7
boto==2.31.0
django-ses==0.4.1

# Misc
gitpython==0.3.2.RC1
raven==5.0.0

# PIL
# django-shorturls==1.0.1
# suds==0.4
# django-celery==3.0.11

django-redis-sessions-fork==0.5.2
django-lionheart-helpers==1.1.12
django-statictastic==0.6.2

django-localflavor==1.0

# Heroku-specific
dj-database-url

gunicorn==18.0
gevent
